combined minio storage and repo
