+++ api/domain/imagemanager/image.go
@@ -1,0 +1,9 @@
+package imagemanager
+
+import "io"
+
+type ImageV1 struct {
+	Meta *Meta `json:"Meta"`
+	URI  string     `json:"url"`
+	File io.Reader  `json:"File"`
+}
